+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Avoid a keyword-argument dispatch (`Base.kwerr` dead branch) in the particle-initialization kernel that triggered AMDGPU hostcalls [#TODO](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/TODO)
 - Enable HIP-graphs acceleration of the batched Fourier transform for AMDGPU, mirroring the existing CUDA-graphs implementation [#1218](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1218)
 - Reformulate trigonometric functions to avoid hostcalls on AMDGPU - [#1210](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1210)